Move shared test helper outside of test
