+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/__tests__/serializeTSComponent-DocUrl.test.ts
@@ -2,7 +2,7 @@ import { Warned } from '../../../../../../common/warning/Warned';
 import { ComponentImplementationInfo } from '../../../../../discovery/component/ComponentInfo';
 import { ComponentMetadata } from '../../../ComponentDefinition';
 import { serializeTSComponent } from '../serializeTSComponent';
-import { getImplementation } from './serializeTSComponent.test';
+import { getImplementation } from './utils/getImplementation';
 
 describe('serializeTSComponent-DocUrl', () => {
   describe('functional component', () => {
