+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/__tests__/serializeTSComponent-Wrappers.test.ts
@@ -3,7 +3,7 @@ import { ComponentImplementationInfo } from '../../../../../discovery/component/
 import { ComponentMetadata } from '../../../ComponentDefinition';
 import { ComponentWrapperType } from '../../../wrappers/ComponentWrapper';
 import { serializeTSComponent } from '../serializeTSComponent';
-import { getImplementation } from './serializeTSComponent.test';
+import { getImplementation } from './utils/getImplementation';
 
 describe('serializeTSComponent-Wrappers', () => {
   describe('when class component with valid wrappers declaration is provided', () => {
