+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/__tests__/utils/getImplementation.ts
@@ -1,0 +1,10 @@
+import { getTypeScriptComponentPath } from '../../../../../../../../test/utils/resources/getExampleComponentPath';
+import { ComponentImplementationInfo } from '../../../../../../discovery/component/ComponentInfo';
+
+export function getImplementation(componentName: string): ComponentImplementationInfo {
+  return {
+    framework: 'reactjs',
+    lang: 'typescript',
+    path: getTypeScriptComponentPath(componentName),
+  };
+}
